+++ src/client/main.tsx
@@ -17,7 +17,7 @@
 
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import { MANIFESTS } from "../common/manifests";
+import { MANIFESTS } from "../common/visualization-manifests";
 import { AppSettings, AppSettingsJS } from "../common/models/app-settings/app-settings";
 import { Timekeeper, TimekeeperJS } from "../common/models/timekeeper/timekeeper";
 import { TurniloApplication } from "./applications/turnilo-application/turnilo-application";
