+++ src/common/utils/rules/rules-evaluator.ts
@@ -15,6 +15,6 @@
  * limitations under the License.
  */
 
-import { Resolve } from "../../models/manifest/manifest";
+import { Resolve } from "../../models/visualization-manifest/visualization-manifest";
 
 export type RulesEvaluator<PredicateVars, ActionVars> = (variables: PredicateVars & ActionVars) => Resolve;
