+++ src/common/view-definitions/index.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-import { BAR_CHART_MANIFEST } from "../manifests/bar-chart/bar-chart";
-import { LINE_CHART_MANIFEST } from "../manifests/line-chart/line-chart";
-import { TABLE_MANIFEST } from "../manifests/table/table";
-import { TOTALS_MANIFEST } from "../manifests/totals/totals";
+import { BAR_CHART_MANIFEST } from "../visualization-manifests/bar-chart/bar-chart";
+import { LINE_CHART_MANIFEST } from "../visualization-manifests/line-chart/line-chart";
+import { TABLE_MANIFEST } from "../visualization-manifests/table/table";
+import { TOTALS_MANIFEST } from "../visualization-manifests/totals/totals";
 import { Essence } from "../models/essence/essence";
 import { ViewDefinition2 } from "./version-2/view-definition-2";
 import { ViewDefinitionConverter2 } from "./version-2/view-definition-converter-2";
