+++ src/common/visualization-manifests/heat-map/heat-map.ts
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-import { Manifest, Resolve } from "../../models/manifest/manifest";
 import { MeasureSeries } from "../../models/series/measure-series";
 import { DimensionSort, isSortEmpty, SeriesSort, SortDirection } from "../../models/sort/sort";
 import { Split, SplitType } from "../../models/split/split";
+import { Resolve, VisualizationManifest } from "../../models/visualization-manifest/visualization-manifest";
 import { Predicates } from "../../utils/rules/predicates";
 import { visualizationDependentEvaluatorBuilder } from "../../utils/rules/visualization-dependent-evaluator";
 import { ActionVariables } from "../../utils/rules/visualization-dependent-evaluator";
@@ -113,9 +113,8 @@ const suggestRemovingMeasures = ({ series }: ActionVariables) => [{
   }
 }];
 
-export const HEAT_MAP_MANIFEST = new Manifest(
+export const HEAT_MAP_MANIFEST = new VisualizationManifest(
   "heatmap",
   "Heatmap",
-  rulesEvaluator,
-  "single"
+  rulesEvaluator
 );
